requirement.txt
appnope==0.1.4
asttokens==2.4.1
blinker==1.8.2
click==8.1.7
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
et-xmlfile==1.1.0
executing==2.0.1
Flask==3.0.3
gunicorn==22.0.0
ipykernel==6.29.4
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.1
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mysql-connector-python==8.4.0
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.3
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==26.0.3
setuptools==69.5.1
six==1.16.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.3
tzdata==2024.1
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.43.0